# idna: optimize punycode handling, add amortized allocation API, improve error reporting